+++ src/js/Events/copyToClipboard.js
@@ -5,22 +5,22 @@
  * @return {void}
  */
 export function copyToClipboardFallback( text, debug = null ) {
-    const textArea = document.createElement("textarea");
+    const textArea = document.createElement( 'textarea' );
     textArea.value = text;
-    textArea.style.top = "0";
-    textArea.style.left = "0";
-    textArea.style.position = "fixed";
-    document.body.appendChild(textArea);
+    textArea.style.top = '0';
+    textArea.style.left = '0';
+    textArea.style.position = 'fixed';
+    document.body.appendChild( textArea );
     textArea.focus();
     textArea.select();
 
     try {
-        document.execCommand('copy');
+        document.execCommand( 'copy' );
     } catch ( err ) {
         if ( debug ) debug.error( 'Failed to copy to clipboard', err );
     }
 
-    document.body.removeChild(textArea);
+    document.body.removeChild( textArea );
 }
 
 /**